GCF/python/requirements.txt
google-cloud-retail
